Move `internedPairKey()` to `schema-utils.ts`.
Also generally clean up the schema files.
